refactor: move HCP1200 and DataladHCP1200 and tests to separate sub-package
