+++ junifer/datagrabber/__init__.py
@@ -2,6 +2,7 @@
 
 # Authors: Federico Raimondo <f.raimondo@fz-juelich.de>
 #          Leonard Sasse <l.sasse@fz-juelich.de>
+#          Synchon Mandal <s.mandal@fz-juelich.de>
 # License: AGPL
 
 
@@ -12,5 +13,5 @@ from .pattern import PatternDataGrabber
 from .pattern_datalad import PatternDataladDataGrabber
 
 from .aomic import DataladAOMICID1000, DataladAOMICPIOP1, DataladAOMICPIOP2
-from .hcp import HCP1200, DataladHCP1200
+from .hcp1200 import HCP1200, DataladHCP1200
 from .multiple import MultipleDataGrabber
