+++ junifer/datagrabber/hcp1200/__init__.py
@@ -1,0 +1,7 @@
+"""Provide imports for hcp1200 sub-package."""
+
+# Authors: Synchon Mandal <s.mandal@fz-juelich.de>
+# License: AGPL
+
+from .hcp1200 import HCP1200
+from .datalad_hcp1200 import DataladHCP1200
